Refactor: 가게 상세 페이지 인증 로직 정리 및 auth 유틸 분리
